packages/python-logic/requirements.txt
opencv-python-headless==4.8.0.76
numpy==1.24.3
pytesseract==0.3.10
Pillow==10.0.0
